Added tests to app
