+++ package.json
@@ -8,6 +8,7 @@
     "react": "^15.6.1",
     "react-bootstrap": "^0.31.2",
     "react-dom": "^15.6.1",
+    "react-loading": "^0.1.4",
     "react-notification-system": "^0.2.15",
     "react-router": "^3.0.0",
     "react-scripts": "1.0.11"
@@ -20,12 +21,14 @@
     "flow": "flow"
   },
   "devDependencies": {
+    "enzyme": "^2.9.1",
     "es6-promise": "^4.1.1",
     "eslint": "^4.5.0",
     "eslint-config-airbnb": "^15.1.0",
     "eslint-plugin-import": "^2.7.0",
     "eslint-plugin-jsx-a11y": "^5.1.1",
     "eslint-plugin-react": "^7.3.0",
-    "isomorphic-fetch": "^2.2.1"
+    "isomorphic-fetch": "^2.2.1",
+    "react-test-renderer": "^15.6.1"
   }
 }
