+++ public/index.html
@@ -20,6 +20,7 @@
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
     <link rel="stylesheet" href="../src/styles/css/bootstrap.min.css">
+    <link rel="stylesheet" href="./main.css">
     <title>Tofaanga</title>
   </head>
   <body>
