+++ src/index.js
@@ -4,10 +4,10 @@ import { Router, Route, browserHistory } from 'react-router';
 import './index.css';
 import App from './App';
 import Err404 from './components/Err404';
-import LoginForm from './components/Login/Login';
-import SignupForm from './components/Signup';
+import LoginForm from './components/Authentication/Login';
+import SignupForm from './components/Authentication/Signup';
 import registerServiceWorker from './registerServiceWorker';
-import Bucketlist from './components/bucketlist';
+import Bucketlist from './components/Bucketlist/bucketlist';
 
 ReactDOM.render(<Router history={browserHistory}>
   <Route path="/" component={App} />
