+++ src/styles/css/login.css
@@ -1,3 +1,11 @@
+/*body {*/
+    /*background: #1e5799; !* Old browsers *!*/
+    /*background: -moz-linear-gradient(-45deg, #1e5799 64%, #1e5799 68%, #1e5799 71%, #ffa8cc 100%); !* FF3.6-15 *!*/
+    /*background: -webkit-linear-gradient(-45deg, #1e5799 64%,#1e5799 68%,#1e5799 71%,#ffa8cc 100%); !* Chrome10-25,Safari5.1-6 *!*/
+    /*background: linear-gradient(135deg, #1e5799 64%,#1e5799 68%,#1e5799 71%,#ffa8cc 100%); !* W3C, IE10+, FF16+, Chrome26+, Opera12+, Safari7+ *!*/
+    /*filter: progid:DXImageTransform.Microsoft.gradient( startColorstr='#1e5799', endColorstr='#ffa8cc',GradientType=1 ); !* IE6-9 fallback on horizontal gradient *!*/
+/*}*/
+
 .form-signin
 {
     max-width: 330px;
@@ -72,4 +80,14 @@
 {
     display: block;
     margin-top: 10px;
+}
+
+.card
+{
+    width: 20rem;
+    border: 2px black solid;
+}
+
+.loader{
+    margin: auto;
 }
